refactor: fjern alle index.ts under regler/
Konsumentar importerer no direkte frå kjeldefilene i staden for
via index.ts re-eksportar. ALLE_*-aggregeringane er flytta inn i
stories-filene som er dei einaste konsumentane. validering/index.ts
er omdøypt til validering/valider.ts.
