+++ packages/uttaksplan/src/felles/uttaksplanValidatorer.tsx
@@ -1,14 +1,16 @@
 import { useIntl } from 'react-intl';
 
 import { useUttaksplanData } from '../context/UttaksplanDataContext';
-import { valider } from '../regler/validering';
+import { valider } from '../regler/validering/valider';
 import { UttaksperiodeValidatorer } from '../utils/UttaksperiodeValidatorer';
 import { LeggTilEllerEndrePeriodeFormFormValues } from './LeggTilEllerEndrePeriodeFellesForm';
 
 export {
     lagStillingsprosentValidator as prosentValideringGradering,
+} from '../regler/felt/stillingsprosent';
+export {
     lagSamtidigUttaksprosentValidator as valideringSamtidigUttak,
-} from '../regler/felt';
+} from '../regler/felt/samtidigUttaksprosent';
 
 export const kanMisteDagerVedEndringTilFerie = (
     perioder: Array<{ fom: string; tom: string }>,
